[geombong/issue-tracker#13] Refactor: remove type folder & add type file in each component
