+++ FE/src/components/common/Input/type.ts
