+++ FE/src/components/common/UserProfile/index.tsx
@@ -1,6 +1,6 @@
 import * as S from './style';
 
-import { UserProfileProps } from '@/type/userProfile.type';
+import { UserProfileProps } from '@/components/common/UserProfile/type';
 
 const UserProfile = ({ imgUrl, userId, size }: UserProfileProps) => {
   return <S.UserProfile src={imgUrl} alt={userId} size={size} />;
